Hardcode download URLs
When working on https://github.com/cucumber/cucumber-rails/pull/378, we would sometimes see that the builds would be unable to connect to the Geckodriver. On further investigation, this seems to be caused by the Geckodriver not being downloaded in the first place. I think that this is due to GitHub's API rate limiting certain build machines from accessing their API.

A better approach would be to hardcode the URLs for the assets -- since they're easily predicatable -- and then to bump this gem whenever a new release comes out.